MB-31409: Use standard golang API to check...
...equality of IP addresses.

The cbauth component is used by services to get the credentials
of a service running on a specific host. This operation involves,
when the cluster is configured using IP addresses, a string
based IP address comparison to fetch the credentials. In certain
situations, this string comparison can fail when IPv6 addresses
are involved. The reason being that a compressed IPv6 address can
be represented in a multitude ways.

The following are the changes made:
[1] Use the standard golang API perform an IP address equality
    check when IP addresses are used.

[2] Added a test case for the IPv6 address case.

Change-Id: I2ff71ac6e393d3643507b68cc09bd86c9e10638f
Reviewed-on: http://review.couchbase.org/100202
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>