Return ClientAuthType as part of GetTLSConfig API.
Currently, the client certificate authentication setting is being
returned by an explicit API called 'GetClientCertAuthType'. However,
the client certificate authentication setting is very specific to
the TLS configuration and it makes more sense to return this setting
as part of the output of 'GetTLSConfig' API. This will also reduce
the number of APIs that the client will have to call in order to
re-configure its web-server with the right TLS configuration.

The changes done are the following:
1. The client certificate authentication state, which is passed as
   a string from ns_server, will be transformed into the corresponding
   'tls.ClientAuthType' and stored in the 'TLSConfig' struct.

2. The 'clientCertAuthState' will no longer be stored in the 'credsDB'.

3. The 'GetClientCertAuthType' API has been modified to return the
   client auth type stored in 'TLSConfig' struct. This API will be
   removed once all the clients start using the 'GetTLSConfig' API.

4. The client certificate parsing code has been updated to use the
   client auth type stored in 'TLSConfig' struct.

Change-Id: I0bbc0a86503927df8474ab97534b9698f2449414
Reviewed-on: http://review.couchbase.org/102992
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>