Add RegisterConfigRefreshCallback API in cbauth.
This will be the new API that the clients would use to register a
callback function that will be called when there is a change in
the following configurations:
1. SSL certificates
2. TLS configuration

The callback function will have the following signature:

'type ConfigRefreshCallback func(uint64) error'

The uint64 argument that the callback takes will indicate which
configuration has changed. Each of the configuration is represented
as a flag and if there are multiple changes then the value resulting
from OR'ing the flags will be returned.

There is one flag defined, called CFG_CHANGE_CERTS_TLSCONFIG, to
signify a change in either the certificates or the TLS configuration.

Change-Id: Ia36f9e344595f99800d977b63c4e2b85afb11fc0
Reviewed-on: http://review.couchbase.org/103217
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>