MB-35602: Update saslauthd-port to Go 1.11.6
When attempting to compile saslauthd-port with ThreadSanitizer
enabled and after the recent change to correctly pass TSan flags to
CGO code, compilation fails with the following error:

    # github.com/couchbase/cbauth/cmd/saslauthd-port
    .debug_pubnames: missing DWARF section for relocation target go.info.main.service
    panic: runtime error: invalid memory address or nil pointer dereference
    [signal SIGSEGV: segmentation violation code=0x1 addr=0x18 pc=0x5294ee]

    goroutine 1 [running]:
    cmd/link/internal/ld.relocsym(0xc4204d2000, 0xc424ca9a48)
        /usr/local/go/src/cmd/link/internal/ld/data.go:607 +0x26ae
    cmd/link/internal/ld.(*Link).reloc(0xc4204d2000)
	/usr/local/go/src/cmd/link/internal/ld/data.go:757 +0x128
    cmd/link/internal/ld.Main()
	/usr/local/go/src/cmd/link/internal/ld/main.go:207 +0x949
    main.main()
	/usr/local/go/src/cmd/link/main.go:58 +0xdb

This is the same issue seen in ns_generate_cert
(http://review.couchbase.org/#/c/113507/) - searching for similar
bugs I see a similar error message in
https://github.com/golang/go/issues/21566, which suggests this
is a problem in Golang toolchain. This issue is marked as
'fixed' by https://go-review.googlesource.com/c/go/+/58630/ which
first appears in go 1.10.

As such, upgrading to go 1.11.6 (we don't appear to use go 1.10, and
1.11.6 is the highest 1.11.x bugfix release).

Change-Id: I0fb8aee876f4938d524e2b35a0658822d43869f5
Reviewed-on: http://review.couchbase.org/113577
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>