MB-52197: Introducing new GetUserBuckets API
The API returns which buckets the user can access. The user might only
have permissions for a specific collection in the bucket but is deemed
able to access to the bucket.
API is currently required for query so that given a user name, they know
what bucket we need to use as tenant for metering and throttling purposes.

Corresponding ns_server change,
https://review.couchbase.org/c/ns_server/+/176996

Change-Id: I58075e0b510592f0c5373fe72f93c3a2c3df07f5
Reviewed-on: https://review.couchbase.org/c/cbauth/+/176997
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>