MB-55145: Change GOVERSION to minor release SemVer granularity
Update the GOVERSION in cbauth to be 1.19 instead of 1.18.1 since we
want it to silently upgrade any patch releases, but not minor
versions. Semver tells us that there should not be breaking changes in
patch releases, so this strategy allows us to automatically pick up
security fixes by just rebuilding instead of needing to manually
change the release version.

Change-Id: Ic260dfc1ab1e4464b29f7637258125b3c8dbe944
Reviewed-on: https://review.couchbase.org/c/cbauth/+/185216
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>