MB-55374: return error in cbauth if splitting nodeport fails
In cbauth, there is a place (during initialization) where we
are (potentially) given a host:port combination, and we are attempting
to split it. If this fails, we seem to just return nil, when we could
return an error instead. Uses similar format to other errors and
returns the initial input as well as the exact error message returned.

Change-Id: I5eb239743690839fa865fcdad3f7f0970847cea1
Reviewed-on: https://review.couchbase.org/c/cbauth/+/187314
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>