MB-61202: Add Guardrail Statuses cache
Adds a cache for guardrail statuses that services may need to read
from. Services can detect changes to this config using a registered
ConfigRefreshCallback, and checking for the
CFG_CHANGE_GUARDRAIL_STATUSES code, and fetching the new config with
cbauth.GetGuardrailStatuses().

Only statuses applicable to the service will be sent to cbauth, so no
filtering would be required on the service end. There will be at most
one status given per resource, so if one severity level is reached,
then any lower severities should be assumed to have been met as well.

Disk usage status example:

Resource = "disk_usage"
Severity = "serious" | "critical" | "maximum"

The thresholds for these severity levels are configured with the
/settings/resourceManagement endpoint in ns_server.

If a resource has not crossed any of the configured limits, the status
will not be listed.

Change-Id: I8fc464354df1e07c2ad4ad7ba8901933a7a9be06
Reviewed-on: https://review.couchbase.org/c/cbauth/+/207501
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>