MB-59455: Made UUID changes backwards compatible with 7.2.2.
Previously, the UUID changes would make 7.2.2 invalidate the
db in every case due to the UUID not being included in the db.
This change makes it proceed as expected when the UUID is
not included, and also makes it possible for an external user
to specify the expected cluster-uuid, which should be used for validation by cbauth when supported by the server.

Change-Id: I1f7be7f67c5fd3a08ca1b18654e966b96b832da8
Reviewed-on: https://review.couchbase.org/c/cbauth/+/201998
Tested-by: Brett Lawson <brett19@gmail.com>
Well-Formed: Restriction Checker
Reviewed-by: Artem Stemkovski <artem@couchbase.com>