MB-61006: introduce IsAllowedInternal cbauth
Introduces a new method IsAllowedInternal to be used in places where
they originally called IsAllowed, but do not want this to produce an
audit record. This method checks if particular permissions are allowed
for a particular user or not. This is generally provided s/t when query
wants to iterate the permissions, we won't spam the logs with audit
messages that weren't produced by a user's actions.

Change-Id: I91845c9bef9d98c83b045c3a41e724071c40b4b2
Reviewed-on: https://review.couchbase.org/c/cbauth/+/207276
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>