MB-63344: Revert "MB-63017 Don't assume loopback addr matches host"
This reverts commit a00115ad91b95d449f50fe2124f9b91487d44443.

Reason for revert: The use case for this behaviour was rediscovered.

A Go service using the SDK gets its node IPs from
/pools/default/nodeServices.
This provides a blank IP for localhost nodes, which we will see for a
cluster_run. The SDK assigns the IP that it contacted the node on as
the IP for those blank nodes. The Go service will then fetch the
memcached credentials through cbauth, requesting the credentials for
that IP. If we don't assume loopback nodes are any host, then we can't
match up that host to the node.

(The tests added in the original patch are being retained, since they
are still useful)

Change-Id: I198928176aafc5558f1eae261b31ece2524c703b
Reviewed-on: https://review.couchbase.org/c/cbauth/+/215186
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>