MB-64921 Add 'hybrid' client cert auth setting
This adds 'hybrid' as a possible setting for ClientAuthState. This
indicates to the service to authenticate internal communication with
client certs when node-to-node encryption is enabled (all/strict).
Otherwise basic auth should be used.

For external clients, hybrid should work the same as enabled.

The services can determine if 'hybrid' or 'mandatory' has been set
by doing:

  TLSSettings, err = cbauth.GetTLSConfig()
  if err != nil {
        // fail appropriately
  }
  if TLSSettings.ShouldClientsUseClientCert {
        // Must specify client cert
  }

This change is dependent on ns_server change:
    https://review.couchbase.org/c/ns_server/+/222550

Change-Id: Ied32eae90dc5d15b953b54390bf98ac6f3b9e7f7
Reviewed-on: https://review.couchbase.org/c/cbauth/+/222547
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>